+++ ahcli.py
@@ -1,32 +1,9 @@
 import urllib.request, urllib.error
 import json
-import csv
+import click
+
 from validators import validate
-
-def converToCsv(path, fileName, data):
-    file = './' + path + '/' + fileName + '.csv'
-    with open(file, 'w') as articles:
-        print(type(data))
-        data = json.loads(data)
-        column_items = data[0]
-        fieldNames = []
-        for key, value in column_items.items():
-            fieldNames.append(key)
-        theWriter = csv.DictWriter(articles, fieldnames=fieldNames)
-
-        theWriter.writeheader()
-        
-        all_items = {}
-        for article in data:
-            for key, value in article.items():
-                all_items.update({key: value})
-            theWriter.writerow(all_items)
-
-def writeToJson(path, fileName, data):
-    file = './' + path + '/' + fileName + '.json'
-    with open(file, 'w') as articles:
-        data = json.loads(data)
-        json.dump(data, articles, indent=4, sort_keys=True)
+from conversions import Conversions
 
 command = input("list / view : ")
 
@@ -39,12 +16,12 @@ try:
     if com[1] != None:
         data = json.loads(the_page)[com[1]]
         article_res = json.dumps(data, indent=4, sort_keys=True)
-        writeToJson('./', 'articles', article_res)
+        Conversions.writeToJson('./', 'articles', article_res)
+        Conversions.converToCsv('./', 'articles', article_res)
     else:
         data = json.loads(the_page)
         res = json.dumps(data, indent=4, sort_keys=True)
-        writeToJson('./', 'articles', res)
-        converToCsv('./', 'articles', res)
+        Conversions.writeToJson('./', 'articles', res)
 except urllib.error.HTTPError as e:
     # Return code error (e.g. 404, 501, ...)
     # ...
