+++ articles.json
@@ -20,7 +20,7 @@
         ],
         "title": "What is love, baby don't heard me",
         "updatedAt": "2019-04-26T11:48:53.196764Z",
-        "views": 4
+        "views": 8
     },
     {
         "author": 6,
@@ -43,7 +43,7 @@
         ],
         "title": "new",
         "updatedAt": "2019-04-26T11:36:42.556970Z",
-        "views": 70
+        "views": 223
     },
     {
         "author": 2,
@@ -66,6 +66,6 @@
         ],
         "title": "new",
         "updatedAt": "2019-04-26T11:24:20.240959Z",
-        "views": 19
+        "views": 24
     }
 ]
