+++ requirements.txt
@@ -1,0 +1,1 @@
+Click==7.0
